[ADD] Complete ECS, needs testing

+++ src/Engine/EntityManager.cpp
@@ -11,7 +11,7 @@ EntityId EntityManager::Create()
         return MakeEntity(index, m_vRecords[index].gen);
     }
 
-    uint32 index = m_vRecords.size();
+    uint32 index = (uint32)m_vRecords.size();
     m_vRecords.push_back({nullptr, 0, 0}); // Row is set to 0 until components are stored in the archetype
     return MakeEntity(index, m_vRecords.back().gen);
 }
